--- plugin/hpsa/__init__.py
@@ -1,1 +1,0 @@
-from lsm.plugin.hpsa.hpsa import SmartArray
