+++ plugin/hpsa_plugin/__init__.py
@@ -1,0 +1,1 @@
+from hpsa_plugin.hpsa import SmartArray
